feat(database): remove the modal when converting blocks to a table (#8275)
fix: BS-1395
